fix compiler warning

+++ lib/select.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: select.c,v 1.37 2007-03-29 00:11:55 yangtse Exp $
+ * $Id: select.c,v 1.38 2007-04-02 02:13:26 yangtse Exp $
  ***************************************************************************/
 
 #include "setup.h"
@@ -479,7 +479,7 @@ int Curl_select(int nfds,
       SET_SOCKERRNO(EINVAL);
       return -1;
     }
-    timeout_ms = (timeout->tv_sec * 1000) + (timeout->tv_usec / 1000);
+    timeout_ms = (int)(timeout->tv_sec * 1000) + (int)(timeout->tv_usec / 1000);
   }
   else {
     timeout_ms = -1;
